ServerlessDocs/hello_world/requirements.txt
requests
fastapi
mangum
pydantic
starlette
python-jose
boto3
